Initial refactor

+++ include/cactus_rt/linux/sched_ext.h
@@ -24,16 +24,11 @@ struct sched_attr {
   uint64_t sched_period;
 };
 
-inline int sched_setattr(pid_t                    pid,
-                         const struct sched_attr *attr,
-                         unsigned int             flags) {
+inline long sched_setattr(pid_t pid, const struct sched_attr *attr, unsigned int flags) {
   return syscall(SYS_sched_setattr, pid, attr, flags);
 }
 
-inline int sched_getattr(pid_t              pid,
-                         struct sched_attr *attr,
-                         unsigned int       size,
-                         unsigned int       flags) {
+inline long sched_getattr(pid_t pid, struct sched_attr *attr, unsigned int size, unsigned int flags) {
   return syscall(SYS_sched_getattr, pid, attr, size, flags);
 }
